requirements.txt
streamlit
PYPDF2
google-generativeai